Send logs from nats to syslog
Redirect the stdout and stderr in the ctl script into
syslog following the standard vcap pattern, in addition to files.

Nats will print to stdout if no log option is specified[1], so this
commit removes the `-l` option passed to nats.

We strip the date from the logs sent to logger using sed, as syslog
will add its own timestamp.

We run `logger` and `tee` with the vcap user by calling `chpst`

We do the redirection at the script level to be consistent with
other services in the CF release. Also, the current gnatsd version
does not support sending logs to syslog and we want to override the
tag name.

[1] http://nats.io/documentation/server/gnatsd-logging/