update NATS server to v1.4.1
The current latest version of the NATS server is v2+ but there are
backwards incompatible changes to the routing protocol which make that
upgrade a little more involved. This version (1.4.1) is the latest
release on the v1.x line which has no breaking changes.

Empirially we've seen that this version fixes some issues with message
loss which was causing some strife with our attempts to benchmark the
system.