add spec tests for nats.conf.erb
We are backfilling some test so we can make our changes with more
confidence that we don't break the conf templating.

We're sorry about the multi-line conf test. The NATS conf files are not
in JSON or YAML or any common format. We would like to find a better
option here.

Co-authored-by: Amelia Downs <adowns@vmware.com>